Move WaitStrategy tests to the right package
And stop testing deprecated methods
